Fix Visual C++ warnings and the extra semicolons of vibes.h

+++ src/core/domains/interval/codac2_Interval_impl.h
@@ -563,7 +563,7 @@ namespace codac2
 
   inline Interval operator""_i(long double x)
   {
-    return Interval(x);
+    return Interval(static_cast<double>(x));
   }
 
   inline double prev_float(double x)
